+++ tests/provider/dwd/mosmix/test_api_stations.py
@@ -85,6 +85,37 @@ def test_dwd_mosmix_stations_filtered(default_settings, mosmix_stations_schema):
     stations = DwdMosmixRequest(parameter="large", mosmix_type="large", settings=default_settings)
     given_df = stations.all().df
     assert not given_df.is_empty()
+    assert given_df.select(pl.all().max()).to_dicts()[0] == {
+        "from_date": None,
+        "height": 4670.0,
+        "icao_id": "ZYTX",
+        "latitude": 79.98,
+        "longitude": 179.33,
+        "name": "ZWOENITZ",
+        "state": None,
+        "station_id": "Z949",
+        "to_date": None,
+    }
+    assert given_df.select(pl.all().min()).to_dicts()[0] == {
+        "from_date": None,
+        "height": -350.0,
+        "icao_id": "AFDU",
+        "latitude": -75.45,
+        "longitude": -176.17,
+        "name": "16N55W",
+        "state": None,
+        "station_id": "01001",
+        "to_date": None,
+    }
+    station_names_sorted = given_df.sort(pl.col("name").str.n_chars()).get_column("name").to_list()
+    assert station_names_sorted[:5] == ["ELM", "PAU", "SAL", "AUE", "HOF"]
+    assert station_names_sorted[-5:] == [
+        "KARLSRUHE/BADEN-BADE",
+        "FELDBERG/MECKLENBURG",
+        "HINTERHORNBACH/OESTE",
+        "UNTERGSCHWEND/OESTER",
+        "ACHENKIRCH/OESTERREI",
+    ]
     # Filter dataframe.
     given_df = given_df.filter(pl.col(Columns.STATION_ID.value).is_in(["01001", "72306", "83891", "94767"]))
     # Verify content of filtered dataframe.