+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - Remove direct tzdata dependency
+- Replace pandas read_fwf calls by polars substitutes
 
 0.63.0 (08.10.2023)
 *******************